Change ApplicationBase
