+++ Packages/UGF.Application/Runtime/ApplicationConfiguredLauncher.cs
@@ -1,0 +1,10 @@
+namespace UGF.Application.Runtime
+{
+    public class ApplicationConfiguredLauncher : ApplicationOrderedLauncher
+    {
+        protected override IApplication OnCreateApplication(IApplicationResources resources)
+        {
+            return new ApplicationConfigured(resources, UseReverseModulesUninitialization);
+        }
+    }
+}
