--- Packages/UGF.Application/Runtime/ApplicationUnity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 251fe3d402bd7384e86c852f5269502e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
