--- Packages/UGF.Application/Runtime/ApplicationUnityLauncher.Editor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 72c6be66f7c84c69b5a6006c9b608d82
-timeCreated: 1603042835
