--- Packages/UGF.Application/Runtime/ApplicationUnityLauncher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 89912ec4ea11aa34786836a25685099b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
